+++ BaseGameDEV/MapEnd.cs
@@ -1,0 +1,7 @@
+namespace MapEnd
+{    class Map
+        {
+            //MapInformation I'd assume
+
+        }
+}
