+++ BaseGameDEV/Program.cs
@@ -6,66 +6,9 @@ using System.Dynamic;
 using System.Net;
 using System.Reflection;
 using System.Security.Principal;
-namespace BaseGameDEV
-{  
-    class CurrentMap
-    {
-        //MapInformation I'd assume
-    }
-    
-    class Player
-    {
-        private string Name;
-        public int[] Pos = [0, 0];
-        public String get() {
-            return Name;
-        }
-        public void set(String iname) {
-            Name = iname;
-        }
-    }
-
-    class PlayerAction
-    {
-        //basic movement so far
-        //can only return strings
-        //Uppercase method ONLY
-        private Player P;
-        public PlayerAction(Player iplayer){
-            P = iplayer;
-        }
-        public void SETNAME(String iname){
-            P.set(iname);
-        }
-        public String GETNAME(){
-            return P.get();
-        }
-        public void UP(){
-            P.Pos[1]++;
-        }
-        public void DOWN(){
-            P.Pos[1]--;
-        }
-        public void LEFT(){
-            P.Pos[0]--;
-        }
-        public void RIGHT(){
-            P.Pos[0]++;
-        }
-        public String GETPOS(){
-            return "(" + (P.Pos[0]).ToString() + "," + (P.Pos[1]).ToString() + ")";
-        }  
-
-        // Test Methods
-        public String ECHO(string subject){
-            return subject;
-        }
-
-        public String SUM(int a, int b, int c) {
-            return $"The sum is: {a + b + c}";
-        }
-
-    }
+using PlayerEND;
+namespace ProgramCompiler
+{    
     class UI
     {
         //Add UI to handle command input and avoids prolong else if statement
